[level 1] Title: 두 정수 사이의 합, Time: 12.37 ms, Memory: 83.4 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.java
@@ -4,10 +4,10 @@ class Solution {
         if(a == b){
             return (long) a;
         }else {
-            int max = Math.max(a, b);
-            int min = Math.min(a, b);
+            long max = Math.max(a, b);
+            long min = Math.min(a, b);
             
-            for(int num = min; num <= max; num++){
+            for(long num = min; num <= max; num++){
                 sum += num;
             }
         }
